requirements.txt
bokeh>=3.5.2
nltk>=3.9.1
SciPy>=1.14.1
hydra-core>=1.3.2
# numpy>=2.1.1